+++ src/core/base/decoder/image_decoder.ml
@@ -28,8 +28,9 @@ let log = Log.make ["image"; "decoder"]
 
 (** Function to retrieve width an height from parameters. *)
 let wh iw ih w h =
-  let frame_w = Lazy.force Frame.video_width in
-  let frame_h = Lazy.force Frame.video_height in
+  let frame_w, frame_h = Frame.video_dimensions () in
+  let frame_w = Lazy.force frame_w in
+  let frame_h = Lazy.force frame_h in
   match (w, h) with
     | None, None ->
         (* By default resize anamorphically to the maximum size wrt the frame *)
@@ -48,8 +49,9 @@ let wh iw ih w h =
     | Some w, Some h -> (w, h)
 
 let wh_string iw ih w h =
-  let frame_w = Lazy.force Frame.video_width in
-  let frame_h = Lazy.force Frame.video_height in
+  let frame_w, frame_h = Frame.video_dimensions () in
+  let frame_w = Lazy.force frame_w in
+  let frame_h = Lazy.force frame_h in
   let f d i l =
     if l = "" then None
     else if l.[String.length l - 1] = '%' then (
@@ -67,8 +69,9 @@ let wh_string iw ih w h =
 (* TODO: negative used to mean from right but I don't think it's a good idea
    anymore (for instance to have a scrolling image). *)
 let off_string iw ih ox oy =
-  let frame_w = Lazy.force Frame.video_width in
-  let frame_h = Lazy.force Frame.video_height in
+  let frame_w, frame_h = Frame.video_dimensions () in
+  let frame_w = Lazy.force frame_w in
+  let frame_h = Lazy.force frame_h in
   let f d frame l =
     if l = "" then d
     else if l.[String.length l - 1] = '%' then (