+++ src/core/optionals/ogg/liq_ogg_decoder.ml
@@ -53,9 +53,10 @@ let converter () =
 (** Convert a video frame to YUV *)
 let video_convert scale =
   let converter = converter () in
+  let video_width, video_height = Frame.video_dimensions () in
   fun buf ->
-    let width = Lazy.force Frame.video_width in
-    let height = Lazy.force Frame.video_height in
+    let width = Lazy.force video_width in
+    let height = Lazy.force video_height in
     if buf.Ogg_decoder.format <> Ogg_decoder.Yuvj_420 then (
       let img =
         Image.YUV420.make buf.Ogg_decoder.frame_width